requirements.txt
asgiref==3.5.2
async-generator==1.10
attrs==22.1.0
# backports.zoneinfo==0.2.1
certifi==2022.6.15
cffi==1.15.1
cryptography==37.0.4
cycler==0.11.0
Django>=3.2.0
fonttools==4.34.4
h11==0.13.0
idna==3.3
jieba==0.42.1
kiwisolver==1.4.4
#matplotlib==3.5.2
mysqlclient==2.2.4
numpy>=1.21.0
outcome==1.2.0
packaging==21.3
#pandas==1.4.3
Pillow>=8.3.1
pycparser==2.21
PyMySQL==1.0.2
pyOpenSSL==22.0.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.1
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.1
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
sqlparse==0.4.2
snownlp==0.12.3
tqdm==4.66.4
trio==0.21.0
trio-websocket==0.9.2
tzdata==2022.1
urllib3==1.26.11
#wordcloud==1.8.2.2
wsproto==1.1.0
requests==2.31.0
lxml==4.9.3

# 深度学习相关
torch>=1.9.0
torchvision>=0.10.0

# Django相关
django-cors-headers>=3.10.0